Moved message handlers into one file
